requirements.txt
aiohttp>=3.10.11
python-dotenv[cli]~=1.0.0
gitpython==3.1.43
json-repair==0.29.2
idna==3.7
jinja2>=3.1.5
Pillow==10.3.0
pyarrow~=16.0.0
pydantic==2.9.1
langchain~=0.3.7
langchain-core~=0.3.35
langchain-community==0.3.17
langchain-google-genai==2.0.6
langchain-cohere==0.3.3
langchain-together==0.3.0
langchain-ollama==0.2.1
langchain-openai==0.3.3
streamlit~=1.38.0
python-pptx~=1.0.2
json5~=0.9.14
requests~=2.32.3
transformers>=4.48.0
lxml~=4.9.3
tqdm~=4.66.5
numpy
scikit-learn~=1.5.1
certifi==2024.8.30
urllib3==2.2.3
anyio==4.4.0
httpx~=0.27.2
huggingface-hub~=0.24.5
ollama~=0.4.7
pypdf
chromadb